+++ Source/Tests/static_mesh_test.cpp
@@ -56,7 +56,7 @@ namespace
         return v;
     }
 
-    std::array<std::uint16_t, kCubeIndexCount> MakeCubeIndices()
+    std::array<std::uint32_t, kCubeIndexCount> MakeCubeIndices()
     {
         return {0, 1, 2, 0, 2, 3, 4, 6, 5, 4, 7, 6, 4, 5, 1, 4, 1, 0,
                 3, 2, 6, 3, 6, 7, 1, 5, 6, 1, 6, 2, 4, 0, 3, 4, 3, 7};
